Support HUGE files in spcomp (bug 5959, r=dvander)
Summary: When there are more than 32768 SLOC in a file, spcomp underflows when storing debug info. These are safe to change as the file-backed struct uses uint32_t already.

Test Plan: Compiled the test plugin from bug 5959. I didn't load it, but it compiled without error and the SMX file is valid.

Reviewers: BAILOPAN

Differential Revision: https://phabricator.alliedmods.net/D2